com32/chain: comments, minor adjustment
- add some comments to clarify c{nul,add,max} modes
- use enums instead of defines
- remove unnecessary integer casts
- avoid casts and uninline guid_is0()
- minor handover comment/flow changes
- clean up some old comment-outs
- add -fno-strict-aliasing as it's not picked anymore by default for
  elf.mk

+++ com32/chain/partiter.c
@@ -765,7 +765,7 @@ struct part_iter *pi_begin(const struct disk_info *di, int stepall)
 	    error("Invalid GPT header's values.\n");
 	    goto bail;
 	}
-	if (!(gptl = disk_read_sectors(di, gpt_loff, (uint8_t)gpt_lcnt))) {
+	if (!(gptl = disk_read_sectors(di, gpt_loff, gpt_lcnt))) {
 	    error("Couldn't read GPT partition list.\n");
 	    goto bail;
 	}
@@ -774,11 +774,11 @@ struct part_iter *pi_begin(const struct disk_info *di, int stepall)
 	    error("WARNING: GPT partition list checksum invalid, trying backup.\n");
 	    free(gptl);
 	    /* secondary array directly precedes secondary header */
-	    if (!(gptl = disk_read_sectors(di, gpth->lba_alt - gpt_lcnt, (uint8_t)gpt_lcnt))) {
+	    if (!(gptl = disk_read_sectors(di, gpth->lba_alt - gpt_lcnt, gpt_lcnt))) {
 		error("Couldn't read backup GPT partition list.\n");
 		goto bail;
 	    }
-	    if (check_crc(gpth->table_chksum, (const uint8_t *)gptl, (unsigned int)gpt_lsiz)) {
+	    if (check_crc(gpth->table_chksum, (const uint8_t *)gptl, gpt_lsiz)) {
 		error("Backup GPT partition list checksum invalid.\n");
 		goto bail;
 	    }
